chore(solver): split call helper modules for lint
